WIP finished without tests 100 coverage

+++ src/contexts/users/handlers/index.ts
@@ -1,13 +1,8 @@
-import create from './create';
-import getById from './get-by-id';
-import getAll from './get-all';
-import deleteById from './delete-by-id';
-import update from './update';
+import create from "./create";
+import getById from "./get-by-id";
+import getAll from "./get-all";
+import deleteById from "./delete-by-id";
+import update from "./update";
+import userAuth from "./user-auth";
 
-export {
-  create,
-  getById,
-  getAll,
-  deleteById,
-  update
-};
+export { create, getById, getAll, deleteById, update, userAuth };
